% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-query-result.R
\name{append_query_result}
\alias{append_query_result}
\title{Combine new results for a query with previously downloaded results}
\usage{
append_query_result(old, new)
}
\arguments{
\item{old}{The \link{query_tbl} of previous results}

\item{new}{The \link{query_tbl} of new results from the server}
}
\value{
A new \link{query_tbl} of the appropriate subclass, depending on whether
the batch is complete.
}
\description{
Combine new results for a query with previously downloaded results
}
\seealso{
\code{\link[=perform_query]{perform_query()}}
}
\keyword{internal}
