% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-tbl.R
\name{as_wiki_tbl}
\alias{as_wiki_tbl}
\title{Representation of Wikipedia data as tibble, with request metadata stored as
attributes}
\usage{
as_wiki_tbl(x, request, continue, batchcomplete)
}
\arguments{
\item{x}{A tibble}

\item{request}{The httr2_request object used to generate the tibble}

\item{continue}{The continue parameter returned by the API}

\item{batchomplete}{The batchcomplete parameter returned by the API}
}
\value{
A tibble: an S3 data.frame with class \code{wiki_tbl}.
}
\description{
Representation of Wikipedia data as tibble, with request metadata stored as
attributes
}
