% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_limit}
\alias{check_limit}
\title{Ensure that the limit is correct for the endpoint. Raise an error if not.}
\usage{
check_limit(limit, max)
}
\arguments{
\item{limit}{The limit to be added to the query}

\item{max}{The maximum allowed for the given endpoint}
}
\value{
\code{limit}, assuming no errors
}
\description{
Ensure that the limit is correct for the endpoint. Raise an error if not.
}
\keyword{internal}
