% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_namespace}
\alias{check_namespace}
\title{Ensure namespace arguments are valid}
\usage{
check_namespace(namespace)
}
\arguments{
\item{namespace}{An integer vector of namespace ids, or NULL}
}
\value{
A character vector of namespace, spliced together with a \code{|}, or NULL
}
\description{
Ensure namespace arguments are valid
}
\keyword{internal}
