% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-query-results.R
\name{continue_query}
\alias{continue_query}
\title{Query the Action API continually until a continuation condition no longer
holds.}
\usage{
continue_query(last_result, predicate, max_requests = 1000)
}
\arguments{
\item{last_result}{The query_tbl of results to complete}

\item{predicate}{The while condition. Results will be continually
requested until this evaluates 'false'.}
}
\value{
A query_tbl: an S3 dataframe that is a subclass of tibble::tibble
}
\description{
Query the Action API continually until a continuation condition no longer
holds.
}
\keyword{internal}
