% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-action-request.R
\name{generate_pages_by}
\alias{generate_pages_by}
\title{Specify which pages to generate from the Action API}
\usage{
generate_pages_by(.req, generator, ...)
}
\arguments{
\item{.req}{A httr2_request, e.g. generated by \code{wiki_action_request}}

\item{generator}{The \href{https://www.mediawiki.org/wiki/API:Lists}{type of generator} to return}

\item{...}{Additional parameters to the generator}
}
\value{
An HTTP response: an S3 list with class httr2_request
}
\description{
See \href{https://www.mediawiki.org/wiki/API:Lists}{API:Lists} for available
generators. Most list actions in the MediaWiki API can also be used as
generators. If you use a list action as a generator, then you will receive
the data in the \code{pages} attribute of the response body, rather than under
the name of the generator. You can \code{generate_pages_by} with
\code{get_page_properties} to choose what properties of each page to return.
}
\details{
When selecting a generator to use, or when setting additional
parameters, you must prepend a "g". For example, to generate pages
by the "categorymembers" list, you must pass "gcategorymembers" as the
generator name. Likewise, to set the limit on the number of category
members to return, you need to pass the parameter "gcmlimit"
}
\examples{
# Search for articles about seagulls

resp <- wiki_action_request() \%>\%
  generate_pages_by("gsearch", gsrsearch="seagull") \%>\%
  httr2::req_perform() \%>\%
  httr2::resp_body_json()

resp$query$pages
}
