% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-action-request.R
\name{get_generator_of}
\alias{get_generator_of}
\title{Specify which pages to generate from the Action API}
\usage{
get_generator_of(.req, generator, ...)
}
\arguments{
\item{.req}{A httr2::request, e.g. generated by \code{wiki_action_request}}

\item{generator}{The \href{https://www.mediawiki.org/wiki/API:Lists}{type of generator} to return}

\item{...}{Additional parameters to the generator}
}
\value{
An HTTP response: an S3 list with class httr2_request
}
\description{
See \href{https://www.mediawiki.org/wiki/API:Lists}{API:Lists} for available
generators. Most list actions in the MediaWiki API can also be used as
generators. If you use a list action as a generator, then you will receive
the data in the \code{pages} attribute of the response body, rather than under
the name of the generator. You can \code{get_generator_of} with
\code{get_page_properties} to choose what properties of each page to return.
}
\details{
\strong{NB:} When setting additional parameters for a generator, ensure to prepend
a \code{g} to the parameter name. E.g. when generating members of a category using
\code{gcategorymembers}, you can set the limit for the request by setting the
\code{gcmlimit} parameter
}
\examples{
# Search for articles about seagulls

resp <- wiki_action_request() \%>\%
  get_generator_of("gsearch", srsearch="seagull") \%>\%
  httr2::req_perform() \%>\%
  httr2::resp_body_json()

resp$query$pages
}
