% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-page-data.R
\name{get_page_metadata}
\alias{get_page_metadata}
\alias{get_page_html}
\alias{get_page_summary}
\alias{get_page_related}
\alias{get_page_talk}
\alias{get_page_langlinks}
\title{Get data about pages from their titles}
\usage{
get_page_metadata(title, language = "en")

get_page_html(title, language = "en")

get_page_summary(title, language = "en")

get_page_related(title, language = "en")

get_page_talk(title, language = "en")

get_page_langlinks(title, language = "en")
}
\arguments{
\item{title}{A character vector of page titles.}

\item{language}{A character vector of two-letter language codes, either of length 1 or the same length as \code{title}}
}
\value{
A list, vector or tibble, the same length as \code{title}, with the desired data.
}
\description{
\code{get_page_metadata()} returns \href{https://en.wikipedia.org/api/rest_v1/#/Page\%20content/get_page_title__title_}{metadata about the latest revision of each page}.

\code{get_page_html()} returns \href{https://en.wikipedia.org/api/rest_v1/#/Page\%20content/get_page_html__title_}{the rendered html for each page}.

\code{get_page_summary()} returns \href{https://en.wikipedia.org/api/rest_v1/#/Page\%20content/get_page_summary__title_}{metadata about the latest revision, along with the page description and a summary extracted from the opening paragraph}

\code{get_page_related()} returns summaries for \href{https://en.wikipedia.org/api/rest_v1/#/Page\%20content/getRelatedPages}{20 related pages for each passed page}

\code{get_page_talk()} returns \href{https://en.wikipedia.org/api/rest_v1/#/Talk\%20pages/get_page_talk__title_}{structured talk page content for each title}

\code{get_page_langlinks()} returns \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Get_languages}{interwiki links for each title}
}
\examples{
# Get language links for a known page on English Wikipedia
get_page_langlinks("Charles Harpur")
}
