% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-page-properties.R
\name{get_page_properties}
\alias{get_page_properties}
\title{Find out more about a page or vector of pages}
\usage{
get_page_properties(.x, by = c("pageid", "title", "revid"), properties)
}
\arguments{
\item{.x}{A vector of pageids, titles or revisionids}

\item{by}{The type of information contained in .x}

\item{...}{Name-value pairs of properties you wish to obtain and further
parameters}
}
\value{
A list of the same length as .x, containing the returned properties.
}
\description{
This function will make one API call per page. So to beware of requesting
data about too many pages.
}
