% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr-history-count.R
\name{history_count}
\alias{history_count}
\title{Vector of history count objects returned from the \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Response_schema_2}{Core REST API}}
\usage{
history_count(value, limit, type, request)
}
\arguments{
\item{value}{Integer vector of history counts}

\item{limit}{Boolean vector indicating whether count is at upper limit}

\item{type}{Type of edits requested}

\item{request}{The httr2_request used to retrieve the counts}
}
\value{
An integer vector of history counts, which handles
}
\description{
Vector of history count objects returned from the \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Response_schema_2}{Core REST API}
}
\keyword{data_type}
