% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-request.R
\name{next_batch}
\alias{next_batch}
\title{Get first batch of results for a MediaWiki Action API request}
\usage{
next_batch(.req)
}
\arguments{
\item{.req}{A httr2_request object describing the API query}
}
\value{
A query_tbl: an S3 dataframe that is a subclass of tibble::tibble
}
\description{
Get first batch of results for a MediaWiki Action API request
}
\examples{
# List the first batch of categories that Albert Einstein belongs to
response <- wiki_action_request(
  prop = "categories",
  titles = "Albert Einstein"
  ) \%>\%
  next_batch()
}
