% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr-page-object.R
\name{page_object}
\alias{page_object}
\title{A page object returned from the \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Page_object}{Core REST API}}
\usage{
page_object(x)
}
\arguments{
\item{x}{List of page objects in JSON format from the REST API}
}
\value{
A page object vector
}
\description{
A page object returned from the \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Page_object}{Core REST API}
}
\keyword{data_type}
