% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-request.R
\name{perform_query_once}
\alias{perform_query_once}
\title{Get first batch of results for a MediaWiki Action API request}
\usage{
perform_query_once(.req)
}
\arguments{
\item{.req}{}
}
\value{
A wiki_tbl: an S3 dataframe that is a subclass of tibble::tibble
}
\description{
Get first batch of results for a MediaWiki Action API request
}
\examples{
# List the first batch of categories that Albert Einstein belongs to
response <- wiki_action_request(
  prop = "categories",
  titles = "Albert Einstein"
  ) \%>\%
  perform_query_once()
}
