% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix-params.R
\name{prefix_params}
\alias{prefix_params}
\title{Add required prefix to URL parameters for MediaWiki Action API request}
\usage{
prefix_params(params, prefix)
}
\arguments{
\item{params}{A character vector}

\item{prefix}{A character vector}
}
\value{
A character vector
}
\description{
Add required prefix to URL parameters for MediaWiki Action API request
}
