% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_timestamps}
\alias{process_timestamps}
\title{Convert passed objects into ISO8601 strings for API requests}
\usage{
process_timestamps(...)
}
\arguments{
\item{...}{Dynamic dots: the objects to be coerced}
}
\value{
A named list of ISO strings, the same length as \code{...}
}
\description{
Convert passed objects into ISO8601 strings for API requests
}
\keyword{internal}
