% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-action-request.R
\name{query_generate_pages}
\alias{query_generate_pages}
\title{Specify which pages to generate from the Action API}
\usage{
query_generate_pages(.req, generator, ...)
}
\arguments{
\item{.req}{A httr2_request, e.g. generated by \code{wiki_action_request}}

\item{generator}{The \href{https://www.mediawiki.org/wiki/API:Lists}{type of generator} to return}

\item{...}{Additional parameters to the generator}
}
\value{
An HTTP response: an S3 list with class httr2_request
}
\description{
See \href{https://www.mediawiki.org/wiki/API:Lists}{API:Lists} for available
generators. Most list actions in the MediaWiki API can also be used as
generators. If you use a list action as a generator, then you will receive
the data in the \code{pages} attribute of the response body, rather than under
the name of the generator. You can \code{query_generate_pages} with
\code{query_page_properties} to choose what properties of each page to return.
}
\details{
To set additional parameters to a generator, prepend the parameter with "g".
For instance, to set a limit of 10 to the number of pages returned by the
\code{categorymembers} generator, set the parameter \code{gcmlimit = 10}.
}
\examples{
# Search for articles about seagulls

resp <- wiki_action_request() \%>\%
  query_generate_pages("search", gsrsearch="seagull") \%>\%
  httr2::req_perform() \%>\%
  httr2::resp_body_json()

resp$query$pages
}
