% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore_names}
\alias{restore_names}
\title{Restore names to an oversimplified list}
\usage{
restore_names(simplified_list, old_list)
}
\arguments{
\item{simplified_list}{A list, passed to list_flatten}

\item{old_list}{The list before it was passed}
}
\value{
simplified_list with names included
}
\description{
purrr::list_transpose has the unfortunate side-effect of deleting
the names from a named list if each constituent list has only 1 member.
This occurs if you request only a single page property from the 'pageprops'
query, as an example. Then each item under 'pageprops' will only have a single
value, and the name of the particular property for that value will be deleted
by list_transpose.
}
\keyword{internal}
