% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-request.R
\name{retrieve_all}
\alias{retrieve_all}
\alias{retrieve_all.httr2_request}
\alias{retrieve_all.wiki_tbl}
\title{Continue requesting data from the MediaWiki Action API until there is none
left.}
\usage{
retrieve_all(x)

\method{retrieve_all}{httr2_request}(x)

\method{retrieve_all}{wiki_tbl}(x)
}
\arguments{
\item{x}{An object describing the desired request. Either a wiki_tbl or a
httr2_request}
}
\value{
A wiki_tbl: an S3 dataframe that is a subclass of tibble::tibble
}
\description{
Typically a request to the \href{https://www.mediawiki.org/wiki/API:Main_page}{MediaWiki Action API} will only return a limited
number of results. The maximum results vary by query, but typically is set at
500. If there are more results on the server, the API will return a \href{https://www.mediawiki.org/wiki/API:Continue}{continue object} allowing you to request
further batches of data. This function handles the continue object and
retrieves all the data for the desired query.
}
\section{Methods (by class)}{
\itemize{
\item \code{retrieve_all(httr2_request)}: Get all available data from a constructed request

\item \code{retrieve_all(wiki_tbl)}: Complete a query from an incomplete wiki_tbl

}}
