% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-xml.R
\name{verify_xml_integrity}
\alias{verify_xml_integrity}
\title{Check that a Wikimedia XML file has not been corrupted}
\usage{
verify_xml_integrity(path)
}
\arguments{
\item{path}{The path to the file}
}
\value{
True (invisibly) if successful, otherwise error
}
\description{
The Wikimedia Foundation publishes MD5 checksums for all its database dumps.
This function looks up the published sha1 checksums based on the file name,
then compares them to the locally calcualte has using the \code{openssl} package.
}
