% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-request.R
\name{wiki_action_request}
\alias{wiki_action_request}
\title{Generate request for the \href{https://www.mediawiki.org/wiki/API:Main_page}{MediaWiki Action API}}
\usage{
wiki_action_request(..., action = "query", language = "en")
}
\arguments{
\item{...}{Parameters for the request}

\item{action}{The action to perform, typically 'query'}

\item{language}{The language edition of Wikipedia to request, e.g. 'en' or
'fr'}
}
\value{
A httr2_response
}
\description{
Generate request for the \href{https://www.mediawiki.org/wiki/API:Main_page}{MediaWiki Action API}
}
\examples{
# List the first 10 pages in the category 'Australian historians'

resp <- wiki_action_request(
  list = "categorymembers",
  cmtitle = "Category:Australian_historians",
  cmlimit = 10
)
httr2::resp_body_json(resp)$query$categorymembers

}
