% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-diff.R
\name{wikidiff_tbl}
\alias{wikidiff_tbl}
\title{A data frame representing the differences between two revisions, following
the \href{https://www.mediawiki.org/wiki/Wikidiff2#JSON}{Wikidiff2 schema}}
\usage{
wikidiff_tbl(x, request)
}
\arguments{
\item{x}{The data. A list of diffs to be formatted as a tbl}

\item{request}{The httr2_request used to retrieve the data}
}
\value{
A data.frame of class \link{wikidiff_tbl}, inheriting from \link{wiki_tbl}
}
\description{
A data frame representing the differences between two revisions, following
the \href{https://www.mediawiki.org/wiki/Wikidiff2#JSON}{Wikidiff2 schema}
}
\examples{
#TODO
}
\keyword{data_type}
