% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikkitidy-package.R
\docType{package}
\name{wikkitidy-package}
\alias{wikkitidy}
\alias{wikkitidy-package}
\title{wikkitidy: Tidy Analysis of Wikipedia}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access data from Wikipedia's major APIs using a consistent interface. Query Wikipedia, retrieve data about individual pages, and use web-hosted analysis tools such as XTools or pageviews. Receive responses as into tibbles or simple vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wikihistories.github.io/wikkitidy/}
  \item Report bugs at \url{https://github.com/wikihistories/wikkitidy/issues}
}

}
\author{
\strong{Maintainer}: Michael Falk \email{michaelgfalk@gmail.com} (\href{https://orcid.org/0000-0001-9261-8390}{ORCID}) [copyright holder]

}
\keyword{internal}
