% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikkitidy-example.R
\name{wikkitidy_example}
\alias{wikkitidy_example}
\title{Get path to wikkitidy example}
\usage{
wikkitidy_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
wikkitidy comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
wikkitidy_example()
wikkitidy_example("akan_wiki")
}
