% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-request.R
\name{continue_query}
\alias{continue_query}
\title{Inner loop of \code{retrieve_all}}
\usage{
continue_query(.req, data = NULL, continue = NULL)
}
\arguments{
\item{.req}{A httr2_request object describing the API query}

\item{data}{Data returned from a previous iteration of the query, if any}

\item{continue}{\href{https://www.mediawiki.org/wiki/API:Continue}{Continue}
parameter returned from previous iteration of the query, if any}
}
\value{
A query_tbl: an S3 dataframe that is a subclass of tibble::tibble
}
\description{
This function drives \code{retrieve_all}, by repeatedly appending the continue
parameters returned by the MediaWiki Action API to the request and retrieving
the next batch.
}
\keyword{internal}
