% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-request.R
\name{core_rest_request}
\alias{core_rest_request}
\alias{wikimedia_rest_request}
\title{Build a REST request}
\usage{
core_rest_request(..., language = "en")

wikimedia_rest_request(..., language = "en")
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Components to add to the URL.
These are passed to \link[httr2:req_url]{httr2::req_url_path_append}. If you pass named
arguments, the names will be dropped.}

\item{language}{The two-letter language code for the Wikipedia edition}
}
\value{
A \code{core_rest_request}, an S3 vector that subclasses \code{httr2_request}
(see \link[httr2:request]{httr2::request})
}
\description{
\code{core_request_request()} builds a request for the \href{https://www.mediawiki.org/wiki/API:REST_API}{MediaWiki Core REST API} is the basic REST API
available on all MediaWiki wikis.

\code{wikimedia_rest_request()} builds a request for the \href{https://www.mediawiki.org/wiki/Wikimedia_REST_API}{Wikimedia REST API} is an additional
endpoint just for Wikipedia and other wikis managed by the Wikimedia
Foundation
}
\examples{
# Get the html of the 'Earth' article on English Wikipedia
response <- core_rest_request("page", "Earth", "html") \%>\%
  httr2::req_perform()

response <- wikimedia_rest_request("page", "html", "Earth") \%>\%
  httr2::req_perform()
}
