% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rest-api.R
\name{core_rest_request}
\alias{core_rest_request}
\title{Base request for \href{https://www.mediawiki.org/wiki/API:REST_API}{MediaWiki Core REST API}}
\usage{
core_rest_request(language = "en")
}
\arguments{
\item{language}{The language edition of Wikipedia you wish to access}
}
\description{
The \href{https://www.mediawiki.org/wiki/API:REST_API}{MediaWiki Core REST API} is
the basic REST API available on all MediaWiki wikis.
}
\examples{
# Get the html of the 'Earth' article
response <- core_rest_request() \%>\%
  httr2::req_url_path_append("page", "Earth", "html") \%>\%
  httr2::req_perform()
}
