% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rest-api.R
\name{get_diff}
\alias{get_diff}
\title{Search for insertions, deletions or relocations of text between two versions
of a Wikipedia page}
\usage{
get_diff(from, to, language = "en")
}
\arguments{
\item{from}{Vector of revision ids}

\item{to}{Vector of revision ids}

\item{language}{Two-letter language code for Wikipedia}
}
\value{
A list the same length as \code{from} and \code{to}, comprising
\href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Response_schema_3}{wikidiff2 responses},
represented as R lists
}
\description{
Any two revisions of a Wikipedia page can be compared using the 'diff' tool.
The tool compares the 'from' revision to the 'to' revision, looking for
insertions, deletions or relocations of text. This operation can be performed
in any order, across any span of revisions.
}
\examples{
# Compare revision 847170467 to 851733941 on English Wikipedia
diffs <- get_diff(847170467, 851733941)

# Each wikidiff2 response contains three main keys:
names(diffs[[1]])
# [1] "from" "to" "diff"

# To see the text that has been inserted, deleted, or moved, look in the
# `diff` key, which contains a numbered list of differences
diffs[[1]]$diff[[2]]$text

# In this case, the diff is of type `0`, which means an insertion
diffs[[1]]$diff[[2]]$type

# The function is vectorised, so you can compare multiple pairs of revisions
# in a single call
# See diffs for the last five revisions of the Main Page
revisions <- wiki_action_request() \%>\%
  query_page_properties(
    "revisions", titles="Main_page", rvlimit=5, rvprop="ids", rvdir="older"
  ) \%>\%
  perform_query_once() \%>\%
  tidyr::hoist(revisions, "parentid", "revid")
diffs <- get_diff(from=revisions$parentid, to=revisions$revid)
}
