% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-rest-api.R
\name{get_langlinks}
\alias{get_langlinks}
\title{Find out which other-language wikis have linked pages}
\usage{
get_langlinks(titles, language = "en")
}
\arguments{
\item{titles}{A character vector of page titles}

\item{language}{The language edition of Wikipedia that the titles come from}
}
\value{
A list of arrays of \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Page_language_object}{page language objects},
the same length as the passed list of \code{titles}. Each page language object
has keys: \code{code} and \code{name} describe the language edition in which the
linked page was found; while \code{key} and \code{title} identify the page on the
linked wiki.
}
\description{
All pages passed to the function must be from the same language edition of
Wikipedia. For example, if you have a set of English pages, you can find all
the interwiki links for those pages. If you have a set of pages from
disparate language editions, you can segment the data using \link[dplyr:group_by]{dplyr::group_by}
or similar.
}
\details{
If the function fails to find any interwiki links for the passed title, it
will fail silently.
}
\examples{
# Get language links for a single page on English Wikipedia
get_langlinks("Charles_Harpur")

# Get a tbl of pages from French Wikipedia using the Action API, then
# find their interwiki links
albums_des_beatles <- wiki_action_request(language = "fr") \%>\%
  query_generate_pages("categorymembers", gcmtitle="Catégorie:Album_des_Beatles", gcmtype="page") \%>\%
  query_page_properties("title") \%>\%
  retrieve_all() \%>\%
  dplyr::mutate(links = get_langlinks(title, language="fr"))


}
