% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-page-properties.R
\name{get_page_properties}
\alias{get_page_properties}
\title{Find out more about a page or vector of pages}
\usage{
get_page_properties(.x, by = c("pageid", "title", "revid"), properties, ...)
}
\arguments{
\item{.x}{A vector of pageids, titles or revisionids}

\item{by}{The type of information contained in .x}

\item{properties}{The desired properties; can be either a string delimited by
"|" or a character vector}

\item{...}{Name-value pairs of additional parameters for the particular
properties requested}
}
\value{
A list of the same length as .x, containing the returned properties.
}
\description{
This function will make one API call per page. Beware of requesting
data about many pages!
}
\examples{
# List all members of a category, then retrieve page information
page_ids <- wiki_action_request() \%>\%
query_list_pages("categorymembers", cmtitle="Category:ATP_Cup") \%>\%
  retrieve_all()
page_info <- get_page_properties(page_ids$pageid, by="pageid", "info")
}
