% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki-action-request.R
\name{get_page_properties}
\alias{get_page_properties}
\title{Choose properties to return for pages from the action API}
\usage{
get_page_properties(.req, properties, ...)
}
\arguments{
\item{.req}{A httr2_request, e.g. generated by \code{wiki_action_request}}

\item{properties}{The properties to request, as either a character vector
or a string with properties separated by "|"}

\item{...}{Additional parameters to pass, e.g. to modify what is returned
by the property request}
}
\value{
An HTTP response: an S3 list with class httr2_request
}
\description{
See \href{https://www.mediawiki.org/wiki/API:Properties}{API:Properties} for a list of available properties. Many have
additional parameters to control their behavior, which can be passed
to this function as named arguments.
}
\details{
\code{get_page_properties} can be used in combination with \code{generate_pages_by}
to choose what information to retrieve about the generated pages. If you
already have a list of page ids or titles, you can build a query using
\code{get_page_properties} to collect information about them.
}
\examples{
# Search for articles about seagulls and return their links to other
# language wikis

resp <- wiki_action_request() \%>\%
  generate_pages_by("search", gsrsearch="seagull") \%>\%
  get_page_properties("langlinks") \%>\%
  httr2::req_perform() \%>\%
  httr2::resp_body_json()

resp$query$pages
}
