% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rest-resource.R
\name{get_rest_resource}
\alias{get_rest_resource}
\title{Get resources from one of Wikipedia's \href{https://www.mediawiki.org/wiki/API}{two REST APIs}}
\usage{
get_rest_resource(
  ...,
  language = "en",
  endpoint = c("core", "wikimedia"),
  response_type = c("json", "html")
)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> The URL components of the
desired resources. Names of the arguments are ignored. The function follows
the \href{https://vctrs.r-lib.org/reference/vector_recycling_rules}{tidyverse vector recycling rules}, so all
vectors must have the same length or be of length one.}

\item{language}{Character vector of two-letter language codes}

\item{endpoint}{The desired REST endpoint, either
"\href{https://www.mediawiki.org/wiki/API:REST_API}{core}" or
"\href{https://www.mediawiki.org/wiki/Wikimedia_REST_API}{wikimedia}"}

\item{response_type}{The expected Content-Type of the response. Currently
.  "html" and "json" are supported.}
}
\value{
A list of responses. If \code{response_type} == "json", then the responses
will be simple R lists. If \code{response_type} == "html", then the responses
will \code{xml_document} objects.
}
\description{
This function is intended for developer use. It makes it easy to quickly
generate vectorised calls to the different API endpoints.
}
