% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr-page-language-object.R
\name{page_language_object}
\alias{page_language_object}
\title{A page language object as returned from the \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Page_language_object}{Core REST API}}
\usage{
page_language_object(code, name, key, title)
}
\arguments{
\item{code}{Two-letter language code}

\item{name}{The name of the language}

\item{key}{The URL-friendly title of the page}

\item{title}{The human-readable title of the page}
}
\value{
A vector of page_language_objects.
}
\description{
A page language object as returned from the \href{https://www.mediawiki.org/wiki/API:REST_API/Reference#Page_language_object}{Core REST API}
}
\examples{
# This is the return type of [get_langlinks]:
austen <- get_langlinks("Jane Austen")
class(austen)
}
\keyword{data_type}
