% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-category-members.R
\name{query_category_members}
\alias{query_category_members}
\title{Build a query to enumerate the members of a category in Wikipedia}
\usage{
query_category_members(
  category,
  namespace = NULL,
  type = c("file", "page", "subcat"),
  limit = 10,
  sort = c("sortkey", "timestamp"),
  dir = c("ascending", "descending", "newer", "older"),
  start = NULL,
  end = NULL,
  language = "en"
)
}
\arguments{
\item{category}{The category to start from. Can be named by title or pageid.}

\item{namespace}{Only return category members from the provided namespace}

\item{type}{Alternative to \code{namespace}: the type of category member to return}

\item{limit}{The number to return each batch. Max 500.}

\item{sort}{How to sort the returned category members. 'timestamp' sorts them
by the date they were included in the category; 'sortkey' by the category
member's unique hexadecimal code}

\item{dir}{The direction in which to sort them}

\item{start}{If \code{sort} == 'timestamp', only return category members from
after this date. The argument is parsed by \code{\link[lubridate:as_date]{lubridate::as_date()}}}

\item{end}{If \code{sort} == 'timestamp', only return category members included in
the category from before this date. The argument is parsed by
\code{\link[lubridate:as_date]{lubridate::as_date()}}}

\item{language}{The language edition of Wikipedia to query}
}
\value{
A request object of type \code{generator/query/action_api/httr2_request},
which can be passed to \code{\link[=next_batch]{next_batch()}} or \code{\link[=retrieve_all]{retrieve_all()}}. You can specify
which properties to retrieve for each page using \code{\link[=query_page_properties]{query_page_properties()}}.
}
\description{
This function provides access to the
\href{https://www.mediawiki.org/wiki/API:Categorymembers}{CategoryMembers}
endpoint of the Action API. Categories in Wikipedia contain two main kinds of
'members': subcategories and pages.
}
\examples{
# Get the first 10 pages in 'Category:Physics' on English Wikipedia
physics_members <- query_category_members("Physics") \%>\% next_batch()
physics_members

# Retrieve interwiki links for the first 10 pages in the 'Beatles Albums'
# category on Danish Wikipedia
# TODO: This currently doesn't work, because incomplete batches aren't properly handled
beatles <- query_category_members("Beatles-album", language = "da") \%>\%
  query_page_properties("langlinks") \%>\%
  next_batch()
}
