% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-request.R
\name{retrieve_all}
\alias{retrieve_all}
\alias{retrieve_all.httr2_request}
\alias{retrieve_all.wiki_tbl}
\title{Continue requesting data from the MediaWiki Action API until there is none
left}
\usage{
retrieve_all(x)

\method{retrieve_all}{httr2_request}(x)

\method{retrieve_all}{wiki_tbl}(x)
}
\arguments{
\item{x}{An object describing the desired request. Either a wiki_tbl or a
httr2_request}
}
\value{
A wiki_tbl: an S3 dataframe that is a subclass of tibble::tibble
}
\description{
Typically a request to the \href{https://www.mediawiki.org/wiki/API:Main_page}{MediaWiki Action API} will only return a limited
number of results. The maximum results vary by query, but typically is set at
500. If there are more results on the server, the API will return a \href{https://www.mediawiki.org/wiki/API:Continue}{continue object} allowing you to request
further batches of data. This function handles the continue object and
retrieves all the data for the desired query.
}
\section{Methods (by class)}{
\itemize{
\item \code{retrieve_all(httr2_request)}: Get all available data from a constructed request

\item \code{retrieve_all(wiki_tbl)}: Complete a query from an incomplete wiki_tbl

}}
\examples{
# Create a request object, then repeatedly retrieve all the pages
# NB: the 'uclimit' parameter sets the number of pages retrieved by each
# API call.
jimbos_contributions <- wiki_action_request() \%>\%
  query_list_of("usercontribs", ucuser="Jimbo_Wales", uclimit=500) \%>\%
  retrieve_all()
# To confirm, the resulting tibble has more than 500 rows
nrow(jimbos_contributions)

# Alternatively, you can try out a request using perform_query_once(), and
# if you are happy with the resulting data, retrieve the rest of the results
# from the returned data frame
preview <- wiki_action_request() \%>\%
  query_page_properties("categories", titles="Albert Einstein") \%>\%
  perform_query_once()
print(preview)
all_results <- retrieve_all(preview)
}
