% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikimedia-rest-api.R
\name{wikimedia_rest_request}
\alias{wikimedia_rest_request}
\title{Base request for \href{https://www.mediawiki.org/wiki/Wikimedia_REST_API}{Wikimedia REST API}}
\usage{
wikimedia_rest_request(..., language = "en")
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Components to add to the URL.
These are passed to \link[httr2:req_url]{httr2::req_url_path_append}. If you pass named arguments,
the names will be dropped.}

\item{language}{The two-letter language code for the Wikipedia edition}
}
\value{
A \code{wikimedia_rest_request}, an S3 vector that subclasses
\code{httr2_request} (see \link[httr2:request]{httr2::request})
}
\description{
The \href{https://www.mediawiki.org/wiki/Wikimedia_REST_API}{Wikimedia REST API} is one of
the three main APIs for WikiPedia and its sister projects. This particular API
is unique because its results are cached around the world can be downloaded with
low latency. As a result, it is also the only one of the main APIs that requires
users to throttle their usage.
}
\examples{
# Get the html of the 'Earth' article
response <- wikimedia_rest_request("page", "html", "Earth") \%>\%
  httr2::req_perform()
}
