% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikkitidy-package.R
\docType{package}
\name{wikkitidy-package}
\alias{wikkitidy}
\alias{wikkitidy-package}
\title{wikkitidy: Tidy Analysis of Wikipedia}
\description{
Explore the world's largest knowledge system using the world's most elegant environment for statistical programming. Why not?
}
\author{
\strong{Maintainer}: Michael Falk \email{michaelgfalk@gmail.com} (\href{https://orcid.org/0000-0001-9261-8390}{ORCID})

}
\keyword{internal}
