% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikkitidy-package.R
\docType{package}
\name{wikkitidy-package}
\alias{wikkitidy}
\alias{wikkitidy-package}
\title{wikkitidy: Tidy analysis of Wikipedia in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A wrapper around Wikipedia's major APIs: the MediaWiki Action API, REST API and Wikmedia REST API. These three APIs provide complementary functionality for the analysis of Wikipedia by social scientists and digital humanists. This package provides a tidy interface to all three APIs, along with a range of functions that solve common tasks in Wikipedia research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wikihistories.github.io/wikkitidy/}
  \item Report bugs at \url{https://github.com/wikihistories/wikkitidy/issues}
}

}
\author{
\strong{Maintainer}: Michael Falk \email{michaelgfalk@gmail.com} (\href{https://orcid.org/0000-0001-9261-8390}{ORCID})

}
\keyword{internal}
