% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikkitidy-package.R
\docType{package}
\name{wikkitidy-package}
\alias{wikkitidy}
\alias{wikkitidy-package}
\title{wikkitidy: Tidy Analysis of Wikipedia}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access 'Wikipedia' through the several 'MediaWiki' APIs (\url{https://www.mediawiki.org/wiki/API}), as well as through the 'XTools' API (\url{https://www.mediawiki.org/wiki/XTools/API}). Ensure your API calls are correct, and receive results in tidy tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wikihistories.github.io/wikkitidy/}
  \item \url{https://github.com/wikihistories/wikkitidy}
  \item Report bugs at \url{https://github.com/wikihistories/wikkitidy/issues}
}

}
\author{
\strong{Maintainer}: Michael Falk \email{michaelgfalk@gmail.com} (\href{https://orcid.org/0000-0001-9261-8390}{ORCID}) [copyright holder]

}
\keyword{internal}
